requirements.txt
absl_py
numpy
torch
torchvision